fix(epic7): wrap useSearchParams in Suspense for /projects/new
